# cc-webapp/backend/current_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Casino-Club F2P API",
    "description": "Backend API for Casino-Club F2P gaming platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/actions": {
      "post": {
        "tags": [
          "Game Actions",
          "Game Actions"
        ],
        "summary": "Log Action",
        "operationId": "log_action_api_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionLoggedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/actions/bulk": {
      "post": {
        "tags": [
          "Game Actions",
          "Game Actions"
        ],
        "summary": "Log Actions Bulk",
        "operationId": "log_actions_bulk_api_actions_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkActions"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Log Actions Bulk Api Actions Bulk Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/actions/recent/{user_id}": {
      "get": {
        "tags": [
          "Game Actions",
          "Game Actions"
        ],
        "summary": "Recent Actions",
        "operationId": "recent_actions_api_actions_recent__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActionItem"
                  },
                  "title": "Response Recent Actions Api Actions Recent  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/missions/": {
      "get": {
        "tags": [
          "Events & Missions"
        ],
        "summary": "Get User Missions",
        "description": "Get the list of active missions and the current user's progress.",
        "operationId": "get_user_missions_api_missions__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/kafka/_debug/last": {
      "get": {
        "tags": [
          "Kafka"
        ],
        "summary": "Kafka Last Messages",
        "description": "Return last consumed Kafka messages (debug).",
        "operationId": "kafka_last_messages_api_kafka__debug_last_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActionCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type",
            "examples": [
              "SLOT_SPIN",
              "GACHA_SPIN",
              "LOGIN",
              "REWARD_GRANT"
            ]
          },
          "context": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "client_ts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Ts"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "action_type"
        ],
        "title": "ActionCreate"
      },
      "ActionItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "action_data": {
            "type": "object",
            "title": "Action Data"
          }
        },
        "type": "object",
        "required": [
          "id",
          "action_type",
          "created_at",
          "action_data"
        ],
        "title": "ActionItem"
      },
      "ActionLoggedResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "action_type",
          "created_at"
        ],
        "title": "ActionLoggedResponse"
      },
...
      "BulkActions": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionCreate"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "BulkActions"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/login"
          }
        }
      }
    }
  }
}